+++ cursos/templates/baseCursos.html
