+++ evl/models.py
@@ -211,20 +211,3 @@ class Users(models.Model):
         managed = False
         db_table = 'users'
         unique_together = (('username', 'email', 'cpf'),)
-
-class ContactUs(models.Model):
-    title = models.CharField(max_length=100, blank=False, error_messages={'required': 'Please enter the title'})
-    name = models.CharField(max_length=100, blank=False, error_messages={'required': 'Please enter your name'})
-    email = models.EmailField(max_length=100, blank=False, error_messages={'required': 'Please enter your email'})
-    cpf = models.CharField(max_length=14, blank=False, error_messages={'required': 'Please enter your cpf'})
-    course_id = models.CharField(max_length=100, null=True)
-    course_category_id = models.CharField(max_length=100, null=True)
-    school_initials = models.CharField(max_length=3, blank=False, null=True, error_messages={'required': 'Please enter your school_initials'})
-    type_conversation = models.CharField(max_length=100, blank=True)
-    description = models.CharField(max_length=5000, blank=False, error_messages={'required': 'Please enter a description'})
-    title = models.CharField(max_length=100, blank=False, error_messages={'required': 'Please enter a title'})
-    # date =  models.DateField(default=datetime.today)
-
-    class Meta:
-        managed = False
-        db_table = 'contact_us_messages'
