+++ evl/urls.py
@@ -14,24 +14,20 @@ Including another URLconf
     2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
 """
 from django.conf.urls import url
-from django.conf.urls import include
+from django.urls import path, include
 from . import views
 
 urlpatterns = [
     url(r'^$', views.home, name='home'),
     url(r'^login/$', views.login, name='login'),
-    url(r'^cursos/$', views.cursos, name='cursos'),
-    url(r'^faleConosco/$', views.faleConosco, name='faleConosco'),
-    url(r'^faleConoscoMensagens/$', views.mensagensFaleConosco, name='mensagensFaleConosco'),
     url(r'^cadastro/$', views.cadastro, name='cadastro'),
-    url(r'^meuscursos/', views.meusCursos, name='meusCursos'),
-    url(r'^certificados/', views.certificados, name='certificados'),
-    url(r'^validarCertificado/', views.validarCertificado, name='validarCertificado'),
-    url(r'^comprovantes/', views.comprovantes, name='comprovantes'),
     url(r'^aluno', views.homeAluno, name='homealuno'),
-    url(r'^baseCursos', views.baseCursos, name='baseCursos'),
     url(r'^dashboard', views.dashboard, name='dashboard'),
     url(r'^o/', include('oauth2_provider.urls', namespace='oauth2_provider')),
     url(r'^secret$', views.secret_page, name='secret'),
+    path('', include('cursos.urls')),
+    path('', include('fale_conosco.urls')),
+    path('', include('certificado.urls')),
+    path('', include('comprovante.urls')),
    # url(r'^accounts/login/$', views.loginOAuth, name="loginOAuth")
 ]
