+++ fale_conosco/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class FaleConoscoConfig(AppConfig):
+    name = 'fale_conosco'
